fix: only show loading when both to and from are selected
Currently, when both from and to are empty and a place is selected the searching illustration and loading mode of the button is briefly triggered. I think this can cause confusion for some users, seeing that something happened on the screen, but in fact nothing did other than updating the URL (since nothing is done if either from or to is empty when the page loads).

Related to https://github.com/AtB-AS/kundevendt/issues/15757