fix: reload page manually to refresh estimated calls
Instead of using a `Link` I changed it to a `button` that calls `router.reload()`. This reloads the page and results in reloading the estimated calls list.

I think this is by no means a permanent solution, but rather a temporary solution to the problem we are currently seeing. At a later point we could look into updating just the list and not the whole page, but then we would be fetching data on the client side rather than the server side.

,> [!NOTE]
> I think the `Link` might be working, but if a quay, or multiple, is collapsed it won't be visible that the estimated calls was updated.