fix: fix issue with widget not selecting place (#196)
The issue with the widget not selecting a place was due to the event listener on blur. The blur event was happening before the "combobox-commit" event which is the one that sets the place. A fix for this was to add event listener to the `mousedown` event instead of the `click` event in the combobox code. That was solved with a patch using `patch-package`. Also mentioned in this issue: https://github.com/github/combobox-nav/issues/54

Also fixed an issue that showed "undefined" if selected place did not have a locality.