fix: remove line name validation
Line public codes doesn't have to be numbers, and can in theory include
any text. In the case of Farte, "R1, P5, 601" is given as an example,
which doesn't work with the current validation logic.

There is room for improvement elsewhere in terms of user feedback on
invalid line filers: They aren't applied unless they exist on the org in
journey planner, but no feedback is given to the user when the public
code entered doesn't exist.